requirements.txt
confusable-homoglyphs==3.2.0
Django==2.1.4
django-crispy-forms==1.7.2
django-js-asset==1.1.0
django-mptt==0.9.1
django-nested-admin==3.1.3
django-registration==3.0
django-webpack-loader==0.6.0
djangorestframework==3.9.1
Pillow==5.3.0
python-monkey-business==1.0.0
pytz==2018.7
six==1.12.0
django-robots-txt==0.5